requirements.txt
streamlit
streamlit-authenticator
openai
pinecone-client 
pinecone-plugin-assistant
pyyaml
pymongo